POC application
This application has been changed into two parts, but still builds a
WAR file.  When run, this WAR file exposes two endpoints.  Both
endpoints, /services and /web, expose the names and types of all of
the beans registered in the two ApplicationContexts.  The
AutoReconfigurationBeanFactoryPostProcessor registers a bean called
'testBean' into any BeanFactory it comes in contact with.  If you hit
/services, you'll see the testBean, but in /web you won't.  This is
because the init-param trick only works for the ContextLoaderListener,
not for the DispatcherServlet's application context.